% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosSquaredWeight.R
\name{CosSquaredWeight}
\alias{CosSquaredWeight}
\title{Estimate normalized model weights based on the cosine similarity of each model's predictions}
\usage{
CosSquaredWeight(converged_models, x)
}
\arguments{
\item{converged_models}{A list containing the fitted models for the individual}

\item{x}{A data frame containing the daily milking records for the individual}
}
\value{
A vector containing the normalized model weights
}
\description{
Estimate normalized model weights based on the cosine similarity of each model's predictions
}
\keyword{internal}
