% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLacModelsMetrics.R
\name{GetLacModelsMetrics}
\alias{GetLacModelsMetrics}
\title{Estimate the Akaike information criterion (AIC), Bayeasian information criterion (BIC), root mean square percentage error (RMSPE) and mean squared error (MAE) for the fitted models}
\usage{
GetLacModelsMetrics(converged_models, x, trait)
}
\arguments{
\item{converged_models}{A list containing the fitted models for the individual}

\item{x}{data frame containing the daily milking records for the individual}

\item{trait}{The name of the column containing daily milking records}
}
\value{
A data frame containing the AIC, BIC, RMSPE and MAE for each fitted model for the individual
}
\description{
Estimate the Akaike information criterion (AIC), Bayeasian information criterion (BIC), root mean square percentage error (RMSPE) and mean squared error (MAE) for the fitted models
}
