% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_my.R
\name{imp_my}
\alias{imp_my}
\title{Impute missing daily milk yields using the ensemble created}
\usage{
imp_my(out.ind = NULL, dim = NULL)
}
\arguments{
\item{out.ind}{The list containing the data frames with the daily production records obtained from the LacCurveFit function}

\item{dim}{A vector with the days in milk where the milk yield will be imputed. It can contain observed and missing DIM}
}
\value{
A data frame containing the imputed milk yields for the days in milk informed.
}
\description{
Impute missing daily milk yields using the ensemble created
}
