% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LacData.R
\docType{data}
\name{LacData}
\alias{LacData}
\title{A data frame containing the daily milk yield for 100 individuals up to 210 days in milk}
\format{
A data frame with 21821 rows and 3 variables:
\itemize{
 \item ID: Unique individual code
 \item DIM: Days in milk
 \item DMY: Daily milk yield
}
}
\usage{
data(LacData)
}
\description{
A data frame containing the daily milk yield for 100 individuals up to 210 days in milk
}
\examples{
data(LacData)
}
\keyword{datasets}
\keyword{internal}
