% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pars.R
\docType{data}
\name{model_pars}
\alias{model_pars}
\title{A data frame containing the models included in the EnsembleLacs package that can have the parameters edited}
\format{
A data frame with 42 rows and 2 variables:
\itemize{
 \item Model_acronym: The acronym of the models that can have the parameters edited
 \item Parameters:  The list of parameters included in each model
}
}
\usage{
data(model_pars)
}
\description{
A data frame containing the models included in the EnsembleLacs package that can have the parameters edited
}
\examples{
data(model_pars)
}
\keyword{datasets}
\keyword{internal}
