% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRankRange.R
\name{ModelRankRange}
\alias{ModelRankRange}
\title{Create a line plot that shows the range of the ranks obtained for each model across the individuals}
\usage{
ModelRankRange(LacCurveFit, metric = "AIC_rank")
}
\arguments{
\item{LacCurveFit}{The object obtained from the LacCurveFit function}

\item{metric}{The name of the metric to be use to plot the model´s ranks}
}
\value{
A line plot that shows the range of the ranks obtained for each model across the individuals
}
\description{
Create a line plot that shows the range of the ranks obtained for each model across the individuals
}
