% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_EMOTIONS.R
\docType{data}
\name{models_EMOTIONS}
\alias{models_EMOTIONS}
\title{A data frame containing the daily milk yield for 100 individuals up to 210 days in milk}
\format{
A data frame with 47 rows and 5 variables:
\itemize{
  \item Model: Model description
  \item Model_acronym: Model acronym used in the EMOTIONS functions
  \item Authors: List of authors of each model available in EMOTIONS
  \item Year: Year of publication of each model available in EMOTIONS
  \item Reference: Link to the original publication of each model available in EMOTIONS
}
}
\usage{
data(models_EMOTIONS)
}
\description{
A data frame containing the daily milk yield for 100 individuals up to 210 days in milk
}
\examples{
data(models_EMOTIONS)
}
\keyword{datasets}
\keyword{internal}
