% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMilkLoss.R
\name{PlotMilkLoss}
\alias{PlotMilkLoss}
\title{Plot the actual daily milk daily production and the predicted values highlighting the detected milk loss events}
\usage{
PlotMilkLoss(
  data,
  ID,
  res.milkloss,
  MY_col,
  MY_pred,
  col = c("red", "blue", "darkgreen"),
  id_col = "ID"
)
}
\arguments{
\item{data}{A data frame containing the observed and predicted daily milking records}

\item{ID}{The ID of the individual that will have the daily milking records plotted}

\item{res.milkloss}{The object with the output of milkloss_detect function}

\item{MY_col}{The name of the column containing the observed milk yield}

\item{MY_pred}{The name of the column containing the predicted milk yield}

\item{col}{The colors of the actual, predicted values, and milk loss events. In this order}

\item{id_col}{The name of the column containing the individual IDs}
}
\value{
A plot with the actual daily milk daily production and the predicted values highlighting the detected milk loss events
}
\description{
Plot the actual daily milk daily production and the predicted values highlighting the detected milk loss events
}
