% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeModels.R
\name{RidgeModels}
\alias{RidgeModels}
\title{Visualize the distribution of model ranks across individuals using ridge density plots}
\usage{
RidgeModels(LacCurveFit, metric = "AIC_rank")
}
\arguments{
\item{LacCurveFit}{The object returned by the `LacCurveFit` function}

\item{metric}{The name of the metric used to plot the model ranks}
}
\value{
A ridge density plot showing the distribution of ranks for the models included in the ensemble
}
\description{
The `RidgeModels` function creates ridge density plots to visualize the distribution of model ranks across individuals.
}
