% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosSquaredWeight.R
\name{CosSquaredWeight}
\alias{CosSquaredWeight}
\title{Estimate normalized model`s weights based on the cosine similarity for each model's predictions}
\usage{
CosSquaredWeight(converged_models, x)
}
\arguments{
\item{converged_models}{A list containing the fitted models for the individual}

\item{x}{data frame containing the daily milking records for the individual}
}
\value{
A vector containing the model normalized weight
}
\description{
Estimate normalized model`s weights based on the cosine similarity for each model's predictions
}
