% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeightLac.R
\name{PlotWeightLac}
\alias{PlotWeightLac}
\title{Plot the actual daily milk daily production and the predicted values obtained by the ensemble model}
\usage{
PlotWeightLac(data, ID, trait, metric, dim, col = c("red", "blue"))
}
\arguments{
\item{data}{A data frame containing the daily milking records}

\item{ID}{The name of the column containing the unique IDs of the individuals}

\item{trait}{The name of the column containing daily milking records}

\item{metric}{The name of the strategy used obtained the predicted values through the ensemble model}

\item{dim}{The name of the column containing days in milk records}

\item{col}{The colors of the actual and predicted values}
}
\value{
A plot with the actual and predicted daily milk production across the days in milk
}
\description{
Plot the actual daily milk daily production and the predicted values obtained by the ensemble model
}
