% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeightLac.R
\name{PlotWeightLac}
\alias{PlotWeightLac}
\title{Plot the actual daily milk daily production and the predicted values obtained by the ensemble model}
\usage{
PlotWeightLac(
  data,
  ID,
  trait,
  metric,
  dim,
  col = c("red", "blue"),
  point_size = 2,
  line_size = 1,
  axis_text_size = 15,
  axis_title_size = 15
)
}
\arguments{
\item{data}{The object generated by the LacCurveFit.}

\item{ID}{The ID of the individual that will have the daily milking records plotted}

\item{trait}{The name of the column containing daily milking records.}

\item{metric}{The name of the strategy used obtained the predicted values through the ensemble model.}

\item{dim}{The name of the column containing days in milk records}

\item{col}{The colors of the actual and predicted values.}

\item{point_size}{Numeric value indicating the size of the observed data points in the plot.}

\item{line_size}{Numeric value indicating the thickness of the regression line.}

\item{axis_text_size}{Numeric value defining the font size of the axis tick labels.}

\item{axis_title_size}{Numeric value defining the font size of the axis titles.}
}
\value{
A plot with the actual and predicted daily milk production across the days in milk
}
\description{
Plot the actual daily milk daily production and the predicted values obtained by the ensemble model
}
