% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeModels.R
\name{RidgeModels}
\alias{RidgeModels}
\title{The function RidgeModels visualizes the distribution of model ranks across individuals using ridge density plots}
\usage{
RidgeModels(LacCurveFit, metric = "AIC_rank")
}
\arguments{
\item{LacCurveFit}{The object obtained from the LacCurveFit function}

\item{metric}{The name of the metric to be use to plot the model´s ranks}
}
\value{
A ridge density plots for the models included in the ensemble
}
\description{
The function RidgeModels visualizes the distribution of model ranks across individuals using ridge density plots
}
